refactor: single source of truth for phone validation (#87)
Frontend and backend previously had duplicate implementations of
isValidPhone/normalizePhone/formatPhoneForDisplay. Moved the canonical
logic into packages/infra/src/lambdas/phone.mjs (plain ESM JS with a
sibling .d.mts for types), where it can be shipped as-is by
lambda.Code.fromAsset.

- sanitize.mjs and handlers import from ./phone.mjs at runtime
- packages/shared/src/types.ts re-exports via a cross-package relative
  path, so UI code keeps importing from @scrappr/shared with no change
- Vite resolves the cross-package .mjs import natively and bundles it

Both sides now literally execute the same module — no drift risk.
